Move /[book] page to /[book] and /[book]:[hadithNumber]
